+++ webknossos/webknossos/annotation/annotation_info.py
@@ -28,7 +28,7 @@ class AnnotationInfo:
 
     def download_annotation(self) -> Annotation:
         """Downloads and returns the annotation that is discribed by this AnnotationInfo object"""
-        return Annotation.download(self.id, annotation_type=self.type)
+        return Annotation.download(self.id)
 
     @classmethod
     def _from_generated_response(